requirements.txt
fastapi[all];
sqlalchemy;
sqlalchemy.orm;
uvicorn;
pydantic;
jinja2;
requests;
python-multipart;
starlette-session;
redis;
pika;